// src/types/navigation.ts
import { type LucideIcon } from 'lucide-react'

export interface NavSubItem {
  title: string
  url: string
  icon?: LucideIcon
  newTab?: boolean
  isNew?: boolean
}

export interface NavMainItem {
  title: string
  url: string
  icon?: LucideIcon
  subItems?: NavSubItem[]
  newTab?: boolean
  isNew?: boolean
}

export interface NavGroup {
  id: number
  label?: string
  items: NavMainItem[]
}